Change TotalPrice of inventory items to Double to support larger totals resolution. Add support for partial sets to remove duplication (requires changes to existing INVENTORY.CSV as well for some older items). Addresses Issue #1 partially (will need to continue to fix old data that used the "Values" convention instead of "Partial Set" in desc field.)

+++ StampCollection/AlbumFamily+Ex.swift
@@ -115,7 +115,7 @@ extension AlbumFamily {
     var theTotalPrice: String {
         // will return the total price of theRefs array
         let total = theRefs.reduce(0.0, {
-            $0 + ($1.theTotalPrice.toFloat() ?? 0.0)
+            $0 + ($1.theTotalPrice.toDouble() ?? 0.0)
         })
         return String(format: "%.2f", total)
     }
